License Cleanup
Replacing the slash-commented full license text in each PHP file with a file-level docblock with linked copyright and license information improves file readability for both humans and computers (e.g. IDEs and documenter scripts), while maintaining a reference to the license in the file.   Additionally, removing the duplicate license text from the readme.